rename survey to disconnect-survey
