+++ _inc/client/state/action-types.js
@@ -119,9 +119,12 @@ export const JETPACK_SEARCH_TERM = 'JETPACK_SEARCH_TERM';
 export const JETPACK_SEARCH_FOCUS = 'JETPACK_SEARCH_FOCUS';
 export const JETPACK_SEARCH_BLUR = 'JETPACK_SEARCH_BLUR';
 
-export const JETPACK_MARKETING_SUBMIT_SURVEY = 'JETPACK_MARKETING_SUBMIT_SURVEY';
-export const JETPACK_MARKETING_SUBMIT_SURVEY_SUCCESS = 'JETPACK_MARKETING_SUBMIT_SURVEY_SUCCESS';
-export const JETPACK_MARKETING_SUBMIT_SURVEY_FAIL = 'JETPACK_MARKETING_SUBMIT_SURVEY_FAIL';
+export const JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY =
+	'JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY';
+export const JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY_SUCCESS =
+	'JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY_SUCCESS';
+export const JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY_FAIL =
+	'JETPACK_MARKETING_SUBMIT_DISCONNECT_SURVEY_FAIL';
 
 export const DEV_CARD_DISPLAY = 'DEV_CARD_DISPLAY';
 export const DEV_CARD_HIDE = 'DEV_CARD_HIDE';
