+++ _inc/client/state/reducer.js
@@ -23,7 +23,7 @@ import { reducer as search } from 'state/search/reducer';
 import { reducer as settings } from 'state/settings/reducer';
 import { reducer as siteData } from 'state/site/reducer';
 import { reducer as siteVerify } from 'state/site-verify/reducer';
-import { reducer as survey } from 'state/survey/reducer';
+import { reducer as disconnectSurvey } from 'state/disconnect-survey/reducer';
 import { reducer as trackingSettings } from 'state/tracking/reducer';
 import { reducer as mobile } from 'state/mobile/reducer';
 
@@ -43,7 +43,7 @@ const jetpackReducer = combineReducers( {
 	settings,
 	siteData,
 	siteVerify,
-	survey,
+	disconnectSurvey,
 	trackingSettings,
 	mobile,
 } );
